server/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pycryptodome==3.19.0
requests==2.31.0
